Moved the AVPlayerWrapperDelegate to a seperate file.
